requirements.txt
mysql-connector-python
kafka-python
python-dotenv
